feat: add initial makefile set
In one side, try to get similar developer experience as in the backend.
In the other side, it adds the automation to generate the 'src/Api'
directory. Update README.md to align with the changes; finally pin the
version of the sub-module to a specific version of the API, update
generate-api to use the new location and clean-up not necessary rules
and commands.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>