HMS-2601 fix: PATCH executed on finish
The code to update the data was executed when doing click on Next button
for the page 3, however it is expected this call is triggered when the
Finish button in the page 4 is clicked.

This change fix the above situation as when we do click on the Finish
button the event triggered is the onCloseClick.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>