HMS-2151 fix: remember settings on step-back to token page
If the user steps back to the token page, settings they provided in
the later pages of wizard are forgotten.

Avoid this by updating `onVerify` to return early if it detects that
the registered state is already `'completed'`.  This avoid
overwriting the domain in the wizard state with the "pristine"
domain data from the `VerifyRegistry` component.