HMS-2151 fix: make it possible to accept defaults
The default "title" field comes from the domain record (i.e.
whatever ipa-hcc defaults it to).  But the wizard does not enable
"Next" on the Service Details page until `onChangeTitle` fires.  In
other words, the user *must* change the value before they can
proceed.

It should be possible to accept the default.  Extract "data is OK to
proceed" behaviour to a new function `onUserInputChange`.  Call this
function from `onChangeTitle`, and *also* from `onVerify` when the
domain data is first received.

As a result, the Next button on the Service Details page should be
enabled immediately, allowing the user to accept the defaults.