HMS-1623 refact: application context
Simplify the application context so it is cleaner and easier to add
changes in the future. It adds AppContextProvider tag and refactor
according to it. This new tag assign the setters and the value of the
current states so it is propagated accordingly to the rest of the
application (no more getter callbacks).

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>